refactor(hooks/useCardNumber):getCardNumber 분리
